Move preference store annotations to separate package
