+++ app/src/main/java/com/marverenic/music/data/store/PreferencesStore.java
@@ -2,9 +2,9 @@ package com.marverenic.music.data.store;
 
 import android.media.audiofx.Equalizer;
 
-import com.marverenic.music.utils.BaseTheme;
-import com.marverenic.music.utils.PresetTheme;
-import com.marverenic.music.utils.StartPage;
+import com.marverenic.music.data.annotations.BaseTheme;
+import com.marverenic.music.data.annotations.PresetTheme;
+import com.marverenic.music.data.annotations.StartPage;
 
 public interface PreferencesStore extends ReadOnlyPreferencesStore {
